AI/requirements.txt
# Core API
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
python-multipart==0.0.20
anyio==4.10.0

# LangChain & workflow
langchain-core==0.3.75
langchain-openai==0.3.31
langchain-huggingface==0.3.1
langchain-pinecone==0.2.11
langchain-text-splitters==0.3.9
langgraph==0.2.67

# Vector store & model clients
pinecone==7.3.0
openai==1.101.0

# Document ingestion
pandas==2.3.2
pdfplumber==0.11.7
tqdm==4.67.1

# Evaluation helpers
requests==2.32.5
